+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Changelog
 
+### 12.106.0
+
+* `[knf/united]` Added method `ToOption` with shortcut `O`
+* `[knf/united]` Added method `ToEnvVar` with shortcut `E`
+
 ### 12.105.0
 
 * `[log]` Added JSON output format